requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain==0.1.10
langchain-groq==0.1.1
langchain-community==0.0.29
langchain-experimental==0.0.32
groq==0.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
slack-sdk==3.27.1
sqlparse==0.4.4
pytest==7.4.4